refactor: rename changeAmount&Page.ts
